pkg/prometheus: Remove rule ConfigMap migration
Remove Prometheus rule ConfigMap live migration code introduced in #1399
[1].

[1] https://github.com/coreos/prometheus-operator/pull/1399
