+++ pkg/prometheus/operator.go
@@ -882,12 +882,6 @@ func (c *Operator) sync(key string) error {
 
 	level.Info(c.logger).Log("msg", "sync prometheus", "key", key)
 
-	// TODO: Remove migration with Prometheus Operator v0.21.0
-	err = c.migrateRuleConfigMapsToRuleCRDs(p)
-	if err != nil {
-		return err
-	}
-
 	ruleConfigMapNames, err := c.createOrUpdateRuleConfigMaps(p)
 	if err != nil {
 		return err
